requirements.txt
mlflow==1.26.1
numpy==1.21.3
setuptools==59.5.0
ipython==7.32.0
pytorch-lightning==1.6.1
pandas==1.3.5
scikit-learn==1.0.2
statsmodels==0.13.2
crowd-kit==1.0.0
timm==0.5.4
torchmetrics==0.9.1
wandb==0.12.18
protobuf==3.20.*
plotly
matplotlib
umap-learn
